@@ -8,7 +8,7 @@ buildscript {
     }
 
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.5.0'
+        classpath 'com.android.tools.build:gradle:7.0.2'
     }
 }
 
@@ -22,7 +22,7 @@ rootProject.allprojects {
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 31
 
     defaultConfig {
         minSdkVersion 16
@@ -33,7 +33,7 @@ android {
     }
 
     dependencies {
-        implementation 'androidx.core:core:1.3.1'
-        implementation 'androidx.annotation:annotation:1.1.0'
+        implementation 'androidx.core:core:1.6.0'
+        implementation 'androidx.annotation:annotation:1.2.0'
     }
 }
