@@ -3,6 +3,4 @@ package com.example.example
 import android.os.Bundle
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity : FlutterActivity() {
-
-}
+class MainActivity : FlutterActivity()
