@@ -24,10 +24,10 @@ import java.util.List;
 /** Handles share intent. */
 class Share {
 
-  private Context context;
+  private final Context context;
   private Activity activity;
 
-  private String providerAuthority;
+  private final String providerAuthority;
 
   /**
    * Constructs a Share object. The {@code context} and {@code activity} are used to start the share
@@ -183,8 +183,9 @@ class Share {
   @SuppressWarnings("ResultOfMethodCallIgnored")
   private void clearShareCacheFolder() {
     File folder = getShareCacheFolder();
-    if (folder.exists()) {
-      for (File file : folder.listFiles()) {
+    final File[] files = folder.listFiles();
+    if (folder.exists() && files != null) {
+      for (File file : files) {
         file.delete();
       }
       folder.delete();
@@ -220,21 +221,15 @@ class Share {
   }
 
   private static void copy(File src, File dst) throws IOException {
-    InputStream in = new FileInputStream(src);
-    try {
-      OutputStream out = new FileOutputStream(dst);
-      try {
+    try (InputStream in = new FileInputStream(src)) {
+      try (OutputStream out = new FileOutputStream(dst)) {
         // Transfer bytes from in to out
         byte[] buf = new byte[1024];
         int len;
         while ((len = in.read(buf)) > 0) {
           out.write(buf, 0, len);
         }
-      } finally {
-        out.close();
       }
-    } finally {
-      in.close();
     }
   }
 }
